requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django==3.1.5
djangorestframework==3.12.2
docker==4.4.1
pip-tools==5.5.0
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
websocket-client==0.57.0
